暂存  初步代码

+++ base64.c
@@ -123,7 +123,7 @@ void base64_encode(unsigned char *indata, int inlen, char *outdata, int *outlen,
 
 void base64_encode_str(char *indata, char *outdata, int *outlen, int safe)
 {
-	base64_decode(indata, strlen(indata), outdata, outlen, safe);
+	base64_decode(indata, (int)strlen(indata), outdata, outlen, safe);
 	return;
 }
 
@@ -184,6 +184,6 @@ void base64_decode(char *indata, int inlen, unsigned char *outdata, int *outlen,
 
 void base64_decode_str(char *indata, int inlen, unsigned char *outdata, int *outlen, int safe)
 {
-	base64_decode(indata, strlen(indata), outdata, outlen, safe);
+	base64_decode(indata, (int)strlen(indata), outdata, outlen, safe);
 	return;
 }
